requirements.txt
openai
python-dotenv
requests
pydantic
termcolor
yfinance
wikipedia
rich
